Create climbingtheleaderboard.cpp
This solution doesn't pass the larger problem sets. I am assuming that the distance method is taking n time. Lower bound should probably bee logn and creating the set should be n. If distance were constant this should run in nlogn, but it is likely that it runs in n square.