Update climbingtheleaderboard.cpp
This is a solution that works, from the discussions section. I can't claim this solution as my own. Although I was thinking of similar approaches by iterating over the alice vector from end to begin (highest score to lowest) and then iterate over the scores vector, increasing the rank every time we find a new value and pushing the rank to the results vector plus 1, whenever we find a value in scores that is smaller than the current alice score. I might try to code this approach.